Caching for Map -> reduces FPS by around 14 times
MapEditor Bugs fixed:
- removed blinking when moving mouse,
-  where no changes were appearing during mapping
- Past Screen Size: Map Disappears -> problem with texture sizes

Admin panels snaps to right edge of main form

Fixed tooltips on Item Spawner

Fixed Server Error when resizing map to smaller size

+++ Server/src/modConstants.bas
@@ -53,7 +53,7 @@ Public Const MIN_MAPX As Byte = 25
 Public Const MIN_MAPY As Byte = 20
 
 ' Player constants
-Public Const MAX_LEVEL As Integer = 99
+Public Const MAX_LEVEL As Byte = 99
 Public Const MAX_STAT As Integer = 255
 Public Const MAX_PEOPLE As Byte = 100
 Public Const MAX_GUILD_MEMBERS As Byte = 100
